Working roll detection
Roll detection now works, and it gives a decently accurate reading for the current roll of the device. Could be made customizable by adding a threshold till the roll actually starts movement, as well as that the roll can be used to make smooth turns, instead of directly 100%.

Input Actions Asset is set for the touch controls, can be used for later in the development process as well.

Changelog:
- renamed scene to PhoneTiltingTest
- Using a text to display the current role state
- Modified InputSystem_Actions.inputactions
  - Removed unnecessary controls and added ones needed for testing.
  - Removed unnecessary control maps (joystick, gamepad, xr, etc.), only remaining ones now are touchscreen and mk
- Added the simulation-devices package to simulate more devices.

works on #31 #18
closes #30